Added code to integrate init flow to link Git repo and to call frameworks api
